Updates
CODA_TF_MEX now has a more accurate memory test and several warnings to prevent the user from breaking it with big or high fs signals. UI tidied up, and advanced settings moved to another tab. There is a switch to override memory limits, but this will most likely cause the system to break if used.